Partial support for -b begin_UTC -e end_UTC options
Currently the -b begin_time_UTC and -e end_time_UTC options don't
do anything.
This is a partial support for the following cases.
1) The end_data_row in the tsv chart header has to be -1.
2) the timestamp column has to be specified in col 5 of the same
   row.
Requirement 1 doesn't really need to be a requirement.
The way the 'clipping' is achieved is by limiting the rows used
in charts to the row where the UTC column is >= begin_time_UTC
and <= END_TIME_UTC.
All the data rows are still added to the sheet. Since they aren't
used in a chart then the unused rows appear as text (even though
they are numbers). This excludes the 'unused' rows from formulas
such as 'average()' over the used and unused rows.