Only pass -s option if SUM_FILE is specified
Currently the -s option is always passed to the resp_2_tsv.sh
script even if the arg to -s is null. This causes the next option
passed to resp_2_tsv.sh to get treated as the arg to -s.
This is not desired.
This change tests if '$SUM_FILE' is not null and only passes the
-s $SUM_FILE option if so.