Change ITP metric_out processing to use RPN equations
Use the new rpn.awk code to enable equations in the code which
processes ITP metrics.
There is some duplication still with the old way of doing
equations. I'll probably eventually change the old way to the
new way.
The RPN equations are used in computing the derived metrics if the
'averaging' option is specified. Previously derived metrics were
just dropped in the summary during averaging.