Add add_all_to_summary option to show all stats in summary
Currently only select stats are shown in summary table.
The selected stats can be selected on the awk cmdline. If a stat
isn't selected there then, if it is a computed metric (like mem
bw) then it is still added to the summary table.
There are cases where I want to show the averages for all the
columns of data.
This commit adds a add_all_to_summary option (the the gen_xlsx.sh
-o option_list). If this option is specified then all non-cmdline
and non-metric averages are also added to the summary.