refactor the 'drow_end == -1' code
If the ending row of a table is -1 then the ending row is set to
the first blank row after the beginning row. This makes it
easier to build the hdrs row_beg, col_beg, row_end, col_end
fields.
The drow_end code is duped in 2 places. I pulled the code into
a function to avoid the duplication. Also, I think that there was
an error where, if the was a blank row right after the 'hdrs' row
then the table would be dropped. Now check that the we are beyond
the beginning row of the table before we look for a blank row.