Add 'hostname' to summary sheet
Currently I might have the summary info for 100s of servers on the
sum_all sheet but I don't put the hostname on the sheet.
So if I have a question about the data for a server I won't be
sure what the host name is.
This commit extracts the hostname from the path.
It assumes that there is a pattern like
subdir_a/host-name/subdir_a
So I split the path and look for the same subdir string before
and after the hostname.
I probably should just get the hostname when I collect the perf
data.