make the sum_all sheet the 1st sheet and add title text
Currently the charts sheet is the 1st sheet. Google sheets takes
a while to load the sheet/draw the charts so see if having the 1st
sheet being all text helps.
Also add each chart's title as a text string above the chart.
This lets me find the chart I want before the chart is drawn.