Add 2nd arg as endtime cutoff for events
Currently the script reads all the data and computes the averages
for all the data.
Frequently I only want a subset of the data.
This commit uses the 2nd arg as a cutoff time.
The time is expected to be the relative offset time from the start
of the data.