Add error checking, max background tasks, and adv sum_all
Currently it can be difficult to detect when an error has occured
in the scripts.
Now for (almost) every awk script, or called bash script, an exit
code is checked and the script is stopped.
A cmdline option is added to set the max background tasks allowed.
Doing -B 0 disables the background tasks