Add code to reduce number of endpoints charted
Currently there might be 400+ endpoints in a data collection.
The endpoints have a 'level' in most of the endpoint names like
aaa/bbb where aaa the category of endpoint.
The endpoint data adds 10s of MBs to the excel files.
So look for the '/' or a '--', get the top level category and
summarize to the top levels.
This cuts down the size by about half in some cases.