Add error checking and change pidstat to reduce data
Add code to emit an error code and detect the error.
Also, previously info like pidstat was just for 1 server at a time.
Now I want to roll up 100s of servers.
So I need to reduce the data if possible and average it over the
servers.
The data reduction is the drop the pid from the process name.
Also to drop the aaa/bbb everthing from the slash onward.
For instance, there are kworker/0, kworker/1 etc. Just sum them
all to kworker.
Also add the 'per process %cpu' info to the sum_all sheet.