Allow averaging for tables whose columns don't match up
Currently the way that the data for multiple servers is combined
is to average the data.
This works okay if each server's table has the same number of
columns and rows and each column's data is the same variable.
That is, column 5 is 'LLC-cache-misses' on all 100 servers for the
perf-stat table.
But some data will have different columns headers.
For instance, the pidstat data will vary depending on what is
running on the server.
This change looks at the headers and, if we are averaging over a
group of servers, then a hash table is built with a column for
each process. THe averaging table is maintained across the list
of servers and then the averaged table is printed at the end.