Add sum_file_no_formula option support
The sum_file output from this script uses excel formulas to get
the averages of each column of ITP data.
This is good in that, if you hide rows, the new averages show up
on the sum_file.
But the forumulas are an issue for reading the sum_file and
creating a comparison of sum_files. The formulas don't show the
value until you open the file in excel.
The sum_file_no_formula option tells the script to compute an
average and use that average in the sum_file. This lets me read
the sum_file directly and use the value in the comparision sheet.