Fix out-of-bound ref, add -d desc opt, handle no sum_file
Currently there might be very little idea (other than the chart
title (if there is a title)) where the data in a chart comes from.
This commit adds support for passing in a -d description' string.
If the input file is a.json then the desc string is put on the
first line of the a.json.tsv file and included in the spreadsheet.
This commit also fixes cases where an index for the odata list
might be out of bounds. The issue comes from combining the data
from  multiple json files where the json data might not have the
exact same number of rows or columns.
Also, currently the output processing assumes that a summary file
will be passed in and opened (for appending). This is not always
the case so check before writing to it.