Add support for -d desc option or desc from .txt file
Currently there isn't a lot of info on where data might come from.
THe muttley/curl json data might use a pretty involved query
string.
This option allows for passing in a -d description_string to
describe where/how the data was created. The string (if present)
will be added to the top line of the output .tsv file.
If the -d desc string is not present or "" (zero length) then
a .txt file is looked for and, if present, the 1st line of the
file is used as the desc.
If the input file is a.json, then the txt file name used would be
a.txt.
The string will be double quoted when it is passed to json_2_tsv.py
since it likely contains stuff like ';' and spaces.