clean up 60secs.pid code
Currently 60secs.sh adds the pid of the 60secs.sh script to a
/root/60secs.pid file. The 1st line of the file is this pid.
The pids (of all the background tasks started) are appended to
this file.
This tells you which pids need to be killed if you want to kill
all the 60secs related processes. Usually the cleanest way to
stop 60secs is to stop whatever task 60secs is waiting on but
sometimes that is hard to do... or you might not want to kill
that task... just the 60secs monitoring stuff.
But the file doesn't need to refer to /root/.
Change the file name to ~/60secs.pid.