Add script to handle combining 'variable header' data
Currently the code to combine multple servers data into
one file assumes that the table of data has the same
column headers. This is usually true.
But for some data collections the headers will vary. In
particular for 'pidstat' type data showing cputime by
process name. Some servers will be running different
processes and so the header for a column will vary.
Even if they are running all the same processes if the
processes don't appear in the input file in the same
order then the column headers will vary.
This script doesn't assume the columns headers appear in
the same order.
Also you can either average the results or sum the results.