Add method of stopping jobs running in background
Currently, to speed things up, I run subtasks in the background
unless you enter '-m 0'.
This does speedup processing but sometimes you want to stop the
processing.
This commit adds code to create a job_${job_id}.pid file to the
current dir. job_id starts at 0 and is bumped up for each top
level dir passed to gen_xlsx.sh. The PID of the background task
created by '../gen_xlsx.sh ... &' is put in the job_x.pid file.
To stop the background tasks, copy the job_x.pid to job_x.stop
where x is the job_id.
The code knows its job_id and its PID and it will check for a
job_x.stop file and, if present, check that the pid in the file
matches the pid of the script.
If there is a match the code will exit.
If the python script tsv_2_xlsx.py has already started it runs to
completion.
The job_id is also passed to sys_2_tsv.sh and sys_2_tsv.sh also
checks for job_x.stop file. It doesn't currently check the pid
in the file so we have to be sure to delete the job_x.stop file
after the run.
Also, code is added to handle the infra_cputime.txt data files.
They have special handling because the headers of each input file
can be different. The redo_chart_table.sh script is used to
combine the infra_cputime.txt files.
The tsv_2_xlsx_${job_id}.inp file has each infra_cputime.txt.tsv
file in it. These files have to be removed from the .inp file and
replaced with the one infra_cputime_sum.txt.tsv file.