Add sum_by_server option and get category for sum_all.tsv
Currently redo_chart_table.sh uses the sum option to combine
multiple dirs. This works okay but, for the infra_cputime, it
gives the total cpus used by the infra routines. This is
interesting but hard to compare (since the the number of cpus
will vary depending on how many servers you collected from).
I added the sum_by_server which does the sum but then divides by
number of cpus per server (so we get an average cpus used server).
Also, the combined sum_all sheet doesn't save a category (and
just uses 'itp' generally).
This isn't so good. Now I extract the category from the sum.tsv
sheet and add it to the sum_all sheet.