Add sum_per_server option
Currently the 'sum' method of combining multiple servers data
files is useful but, if we are doing something like summing by
cpus then it can be hard to compare runs with different number
of servers.
The sum_per_server divides the sum by the number of servers
(assuming that there is one input file per server).
Before, the infra_cputime charts would show sums adding up to the
total number of cpus on all the servers. This is interesting but
it makes it hard to compare say 1T (32 cpus/server) vs b19a (48
cpus/server).
The sum_per_server still does a sum but then divides it by the
number of servers. This lets me see then the average per server.