Allow for different ps columns positions
Currently the ps output is assumed to be:
PID STATE TTY CPUTM CMD
but I want to add RSS and VSZ, drop STATE and TTY.
So have to save the postion for each column of data now.
And add a chart for RSS by process and another chart for VSZ by
process.