Make less verbose and add scriptname lineno to echo
The scripts can print lots of info. Now that the default is to run
in the background it can be difficult to identify which script is
printing some info.
For gen_xlsx.sh, I've changed almost all the echo statements to
start with "$0.$LINENO ..." so that the script name and script
line number prefixes each echo output line.