Allow multiple -S options to skip tsv_2_xlsx.py and fix sum.tsv
Currently if you do -S then the sys_2_tsv.sh/tsv_2_xls.py call is
skipped.  There is also a call to tsv_2_xls.py in gen_xlsx.sh.
The idea is that, if you are processing multiple directories, then
you don't want to create an xlsx file in each dir. So the -S option
lets me skip creating an xlsx in each dir.
Rather I want to combine the multiple dir output files into 1 xlsx
and that is what the gen_xlsx.sh/tsv_2_xlsx.py accomplishes.
Sometimes I don't want to generate any xlsx file.
If I do ' -S -S ' then the 2nd tsv_2_xlsx.py call (in gen_xlsx.sh)
is avoided as well.

Lastly, the way that the sum.tsv files are combined can lead to
dropped metrics. The metrics name is hashed and used as an index
for saving the values. The problem is that the column I use in
the sum.tsv file can sometimes have the repeated values.
This change uses both the category column and the metric column
now in the hash.