Change how duplicate metrics are handled
Currently the last column of the compare sheet is the metric. It
is not frequent but you can have duplicate metric strings.
For instance, you might have cputime for java (where 'java' is the
right-most label) and rss memory size for java (where java is
again the right-most label).
To differentiate the labels currently I append a _X to the label.
where X is 1 or more.
Rather than append _X this commit uses a label like 'cputime java'
or 'rss memory size java' to differentiate the labels.