use infra end time if > run_log endtime
The end_tm gotten from the run_log can be too short if the job
is run in the background.
So check if there is an infra cputime file, get the end time
from that file and check if it is > run_log endtm.
Use which ever end_tm is later.