Add code for diskstats and change bar chart
Add some initial data for supporting /proc/diskstats. No chart yet
but the script should parse the data file without errors.
Change the column chart to be row oriented. Putting all the data
in a column make it all the same variable. I like to be able to
select a bar and then delete it (so the columns/axis readjust).
But for the column oriented data then all the data gets deleted.
If I put all the data in a row then each value is a different
series and deleting 1 bar doesn't delete all the data.