Allow select range of dirs and add perf_stat to redo script
Currently you can limit the number of dirs to be processed using
-N number. This limits processing to the 1st number dirs (the
first 'number' of server data dirs).
This is useful to debug issues more quickly (instead of plowing
through 100s of dirs to see if you've fixed the issue.
But some issues only show up for a range of dirs (like the the
4 dirs say).
This commit lets you enter a range of dirs to be processed, like
-N 100,103. This will process dirs 100, 101, 102, 103.
Also change the handling of the perf_stat data.
Currently the data for the different dirs is combined by
tsv_2_xlsx.py. The py script doesn't properly handle tables where
the columns are not in the same order.
This commit calls redo_chart_table.sh to process the tables
instead.

Also changed the search for the beg_time,end_time. I usually get
it from the run.log file which has a beg/end timestamp.
But there was a glitch in my code (at one time) where run_itp.sh
ended early. So I added a check for beg/end time from another file
(if present).