Handle charts with a column list and variable columns
This hdrs line can have a list of columns to be charted. This is
usually (always?) used to refer to the data in another table.
The problem is if we are combining the data from different files
and the columns have different headers. This case isn't handled.
This change saves off the column name for each column reference
and then makes sure that the 'same named columns' are summed
together.