remove dead code, generalize other code
Comment out an early exit (added during debugging) in get_dir_list.
Remove a block of commented out code.
If the perf_stat txt file is present but not valid (just the
header) then we can't extract the begin/end timestamps from it.
So add a check for the run.log file if the perf_stat file is too
few lines.
Currently redo_chart_table.sh gets called (for infra_cputime and
perf_stat) if there is an infra_cputime file.
There needs to be a separate check for the perf_stat file and
only call the redo_chart_table.sh for perf_stat if there is a
perf_stat file.