Handle 'ps -eo fmt_str' for all processes in __date__ section
Currently the __date__ section gets the output of
ps -o fmt_str -C xxx -C xxx2
so this is a small section of the list of processes.
To check my code, I wanted to do ps -eo fmt_str to get all the
running processes.
But this is a lot of processes and a lot of them don't even get
any cputime over the course of the run.
So most of the code changes collapse multiple instances of procs
to 1 proc (like kworker/0 kworker/1 kworker/...num_cpus) to just
kworker).
The plst array shows the processes that get collapsed.
The other part of the change is to omit displaying info for
processes that didn't run at all... that's a lot of processes
sometimes.
This cuts down on the data in the xlsx file. but a 12 hour, 4 sec
sampling interval xlsx is still 30 MBs.