Add code to compute 'SI new score' for NCU benchmark
We use 3 components of SpecInt CPU2017 as an NCU benchmark.
We need to compute a score to check the relative performance of
these tests. Previously we use perf stat to show the relative
performance (via TMAM %retiring, %backend bound, %frontend
bound and %bad speculation). If we don't have perf (such as on
cloud boxes) then we need a 'score'.
Take the lowest 'score' of each component subtest, get geomean
and use that as 'SI new score'.
Other changes are to add an option to not produce an mpstat
%cpu chart for each cpu.
These charts were okay when we didn't put all the charts on one
page but now it is too much info to scroll through.
See option 'mpstat_skip_percpu_charts'.
More changes to handle basically empty perf_stat files (due to
not being able to collect anything on cloud).