add alt begin epoch timestamp, filename for outfile
If the timestamps in the perf_stat txt file aren't in the right
format (not UTC) then the script won't be able to get the initial
epoch timestamp.
Allow for passing in an alterantive begin timestamp.
Also, the output .tsv file is currently just the stdout. So if any
extra debug statements are added then the extra printfs muck up
the output file.
Add an input -v out_file="filename" and printf to the out_file.

Also fix an error where the AMD %not_halted eqn didn't divide by
the interval.