Add out_file option for perf_stat_scatter and specjbb results
Now perf_stat_scatter.sh requires the name of the output tsv file
as an option so pass it.
Currently specjbb max-jOPS and crit-jOPS aren't added to the
summary tsv file (even if specjbb has been run).
So check for the specjbb output info and add max-jops/crit-jops
if they are present.