Add support for phase file subsetting of data
Currently the phase file stuff is not working.
What I'd like to do is make it where I can identify the phase for
the perf data and mark each row in the table with the subtest.
Then we can create a filter on the table and the charts (for that
page) will automatically show only that data.
This will let us compare different runs subtest more easily.
This change implements those ideas.
A side-effect is that the epoch timestamp column (column A) of the
perf table gets overwritten with the phase name (so the epoch ts
is overwritten). Probably a better way would be to add a column
with the phase but the logic is too complicated for now and, as
far as I can tell, the epoch timestamp isn't used by any other
script.