Redo avg+3*stdev stats with pXX percentile stats.
Currently to try and capture variation in a variable I compute the
average and standard deviation. This would work fine if the data
was in a normal distribution... but it isn't.
So redo the avg+3stdev stuff to do p10...p99...p100 stats.
Since we are combining data from many hosts this gets complicated.
I can't compute pxx on the averages from each host.
So I add a 'variable val_arr' line to the summary page of each
host which has all the values of the variable over the run.
Then, when I combine the hosts data, I concat the
'variable val_arr' arrays and to the pxx on these values.
So I compute the pxx stats on all the samples of the variable
over all the hosts.

I also wanted to be able to subset the data that gets combined by
host config (such as SKU type). So I added a -s SKU option which
selects only the directories which have that SKU type.
The hostname has to be part of the path (currently this is how
I fetch results from hosts... by making a subdir with the host
name). And you have to have a file that has the config details
for each hostname. In this case the SKU of the host.
If the sku matches then the directory is used.