Redo stdev stats to be pXX stats
Currently I have average and stdev to characterize variation of a
variable. This doesn't work for the variables collected (near as
I can tell) since the variables aren't normally distributed.
So switch to using percentile stat (pxx stats) instead.
For each host add an array of values to the summary so when we
combine the hosts, we can do percentiles on all the samples on all
the hosts.