Change the way NCU score is calculated
NCU is run on 25%, 50%, 75% and 100% of the cpus.
Get a score/cpu for each level, then average the 4 numbers for
overall score.
Added complication is if you can't do a balanced run at say 25%.
This is the case for 48 cpu boxes. Have to do groups of 4 (4 cpus,
2 cores+HT_extra_thread) with equal load each socket.
So this leads to units of 8 cpus (4 cpus per socket).
But .25*48 = 12 cpu... so we can't do balanced at 25%. So straddle
the 25% with a group of 8 and 16 and then take average of those 2
scores... same for 75% (.75*48 = 36 so do 32 & 40 and average).