Add cpu2017 subtest chart. add impi avgs sum_file
Currenty the no chart is produced to show when the each cpu2017
subtest begins and ends.
Add a chart for this.
Currently only some fields from the ipmi sdr are added to the
sum_file. This commit computes and average for each field and
adds an average of each sensor to the sum_file.