Handle '' data (empty csv data) better and add opt
Previously almost all the input data files were created by 60secs
sys_2_tsv.sh or related scripts using tsv files.
At times it is easier to create csv or semicolon delimited files
and then feed them to tsv_2_xlsx.py to get charts, xlsx files.
Now we need to handle missing values (which will be set to '') in
the csv case.
So handle more conditions where the first value saved might be
'' or maybe the 1st val is numeric and the 2nd is '' then handle
those cases.
Also, for latency charts there are a lot of missing values.
This is for plotting mlc's loaded latency matrix and, for excel,
I don't know how to get multiple x+y series on the same xlsx
chart... I can get 1 X column and lots of Y columns.
And since the loaded latency chart has mem bw as the X axis then,
for multiple SKUs (with possibly no match on X bw values) then
I have to have every bw value (19 per loaded latency test) in the
total X list. This means that most of the Y values will be empty.
The default for charts with empty (missing) values is to to show
a gap. This is not good for the loaded latency chart whhere the
curve will only have a handful of values.
So add a chart option 'chart_show_blanks_as{span}' (passed as an
option -O to tsv_2_xlsx.py). This causes all charts to connect the
nonblank values with lines. 'span' can also be 'gap' (the default
near as I can tell) or 'zero' (plot 0 for missing values).