Allow for better clipping and diff fmt mpstat data
The format of mpstat output changed or maybe the options needed
to get the older format changed. In any case my mpstat awk script
had to be updated to detect and handle the different format.
Also the script collecting mpstat data sometimes will didn't end
when I wanted it to end...so it has too much mpstat data.
Now I look for the run.log file 1 dir above 60secs dir and get
the endtime from that file and clip the data to that timestamp.