Allow for cpu2017 subtest subphases
To debug top_down methodology profiling we need to allow for
breaking out the sub-phases of a subtest. The subphases can be
quite different.
The 60secs code already creates a phase file for each cpu2017
subtest (perlbench, gcc, mcf, etc).
Add code for perlbench's 3 sub-phases. Instead of 500.perlbench_r
appearing in the phase file you'll see
500.perlaaa_1 beg_tm end_tm elap_tm
500.perlbbb_1 beg_tm end_tm elap_tm
500.perlccc_1 beg_tm end_tm elap_tm

Also change the line "if (bm_nm != "" && (copies+0) > 1)" to
allow for running cpu2017 with just one copy.

Lastly, when I create the chart data table from the cpu2017
phase file, the 2nd column is the epoch timestamp.
This is not very helpful. It needs to be a relative timestamp.
So subtract the begin time.