changes to simplify not-averaging files
Currently most of the time I am combining 100s of files from the
fleet of servers.
Another usage model is to run a few jobs to do things like compare
cpu2017 on half the cpus on a server vs running on all the cpus.
In the latter case I want to see all the charts for each run, I
don't want to average the runs together.
This requires lots of changes.
The key is to omit the -A and -a options (averaging options).
There has to be someway to differentiate the file sets from each
other. I probably should have an option to do this but currently
I look for a 'cpus' substring like '_12cpus_' in the dirname and
set the prefix to '12cpus'.
We have to pass this prefix for each fileset down to
tsv_2_xlsx.py as well. And previously I only had 1 prefix for
all the filesets. Now do a per fileset prefix.

If we aren't averaging then the beg/end time clipping isn't used.
I don't have a way (currently) to set a per fileset beg/end time.