Add sleep script for use in scripts
Lots of scripts do sleep and I need to kill the sleep to do an
orderly shutdown of the script. For instance, I might have perf
do a 'sleep 10' but there are many sleep processes on the system.
So I use pfay1_sleep.sh and can do 'pkill -2 pfay1_sleep.sh'