Add do_perlbench_subphase{1} option support
We are interested in the 3 subphases of the cpu2017 perlbench
benchmark. sys_2_tsv.sh looks for the cpu2017.001.log file and
creates a phase file for phase (perlbench, gcc, etc) with a begin
and end timestamp.
Some of the workloads have distinct subphases. perlbench has 3
subphases. Setting this option string 'do_perlbench_subphase{1}'
in the string of options tells the 'create the phase_cpu2017.txt
phase file' code to look for the begin/end of each perlbench
subphase and add each subphase to the phase file.
The subphases are named perlaaaa, perlbbbb, perlcccc... not good
names but the code later assumes that phase names are 8 positions
long.