Add some routines for running cmds, fetching data
install_and_run_on_cloud.sh is my main utility to run cmds on a
list of servers and fetch the output.
It has tons of options. Most of then assume that there is a
DynoInstallFolder (with specint, specjbb, stream, geekbench, fio,
etc) and they run scripts from that dir.
But it is still useful to have this script.
find_closest_run_cmd_match.sh is called when the arg to -r isn't
in the list of available cmds.
Then find_closest_run_cmd_match.sh is called to list the closest
match to the list of available cmds.

gen_report.sh is called with the -r report cmd. It tries to
extract benchmark results and put it into a ; delimited file.
fetch_output.sh is a script which install_and_run_on_cloud.sh uses
to tar.gz up the output dir on each server prior to scp'ing the
output from the server back.