Always regen the phase_cpu2017.txt file
If there is a cpu2017 CPU2017.00*.log file under the dir being
processed then generate a phase_cpu2017.txt file.
Currently the phase file doesn't automatically get regenerated.
This is an issue if we do things that need a new phase file.
Say I want to zoom into each workload with a component.
For instance, perlbench has 3 different workloads and I want to
get the perf stats for each workload.