read sheets_$JOB_ID.txt file for list of file for xlsx
Currently I have a search strategy in sys_2_tsv.sh where I look
for dirs and data files to process.
And I have another strategy in gen_xlsx.sh to find the output
files created in sys_2_tsv.sh.
Most of this change is have sys_2_tsv.sh create a file with the
list of output files for a given input dir.
The list is put in a file named 'sheets_${JOB_ID}.txt' in the
current dir.
Now gen_xlsx.sh has a list of paths passed to  sys_2_tsv.sh.
I have to process the sheets_*.txt files in the same order.
So the first arg on each line of the sheets*.txt file is the path
for the output files (the list of output files are below the dir
which is the 1st arg on the line).
Weird path names can cause an issue for matching up the directory
names. So I get a 'canonical path' by doing:
CAN_DIR=`cd $DIR; pwd`
Doing the 'cd ...; pwd' gets rid of stuff like dbl //'s.
The main point is now I don't have different or duplicate ways
to find the output files from sys_2_tsv.sh