Use the highest CPU2017.00*.log file
I run cpu2017 by hand to run individual subtests.
I do 1 runcpu to build the binaries I want and another runcpu to
actually run the benchmarks.
So the 'build' run is CPU2017.001.log and the actual run is
CPU2017.002.log.
I only want the last CPU2017*.log file.
So check for an 002 file first and then (if there isn't an 002)
then check for 001 file.