Look for CPU2017.002.log then CPU2017.001.log file
Aas mentioned in the previous commit, if tehre are 2 cpu2017 log
files in the dir then we want the highest numbered log file.
The CPU2017.001.log might be the 'build the binaries' run and then
the 002 file is the actual run.
If there is just 1 file then use that file.