Allow for more detailed perf stat output
Currently this script just collects/summarizes at the system level.
It can be useful to run it on just a set of cpus or to collect
per cpu data and report the data by cpu.
Say we want to run taskA on the HT0 thread of each core and run
taskB on the HT1 thread of each core.
This change lets you specify a cpu_list from which to collect the
perf data.
Also adds a -W args_to_perf_stat option to do things like:
-W " -A "  to report out per cpu event counts.
This is useful for the case of one task on HT0 and another on HT1.