Add more debug info and push_row_tmr operator
Add more debug info. Some of the error messages don't have any
identifier as to who is printing them.
Also save errors into bc_err variable so it can be examined in
perf_stat_scatter.awk (and errors handled).
In adding support for 'per-cpu' perf stat data, now I need to
know the interval for an event. For instance, when computing
frequency, I need to know how the time interval over which I'm
accumulating cycles. The perf data reports the nanoseconds for
each measurement. I accumulate the nanoseconds and the cycles for
each measurement.
The 'push_row_tmr' operator pushs the nanoseconds for thata event
onto the stack of the equation.