Add 'perf stat -A' (per cpu) support. Catch bc_eqn errors.
Currently I usually only collect system level perf data. This is
okay usually.
Sometimes it is useful to get per cpu perf stat data. Say we are
running 1 specint subtest on the HT0 (hyper-thread 0) cpus and
another specint subtest on the HT1 cpus.
It is helpful to see how each subtest responds to running with
another subtest.
This requires a little bit of changes to reading the data (see the
cpu_col variable) and a lot of changes to the equations.
Equations that compute a 'per second' metric were changed to use
the accumated nanoseconds for each event. See the push_row_tmr
operator in the equations and in rpn_eqn.awk and bc_eqn.awk.
I'm slowly switching the old way I did equations to the bc_eqn
way.
Also I don't think that some of the latency computations were
right so I fixed those.
I added a way of selecting which perf data to use. It is a crude
way now but it works.
If the options has a string like:
file_sets{/b19a.*perlbench_r/:perf_cpus=0-23}
then this will get passed to perf_stat_scatter.awk and the awk
script will select just cpus 0-23 and ignore the rest of the cpu
perf data.
The file_sets stuff probably will end up being a json file but
for now it is an options string.
The /b19a.*perlbench_r/ lets me select which filename I'm matching
to if the regex matches the filename then the args are passed to
the perf awk scripts.
This gives me a way to match "which cpu perf data to use" based on
"which strings are present in the file set name".
For instance, in the above case, the path has the b19a string and
the perlbench_r string. The b19a system has 48 cpus and I ran
perlbench on the HT0 cpus (cpus 0-23). So if the path matches
the regex then perf is told to just use cpu0-23 for the metrics.
Lastly, the bc_eqn.awk script gets called but it wasn't checking
for errors. Now check for errors and exit if errors found.