Add support for file_sets{} in options
Currently there isn't a way to set options that take effect if the
data for the xlsx is say, for a particular benchmark or say,
running a benchmark A on cpus 0-23 and benchmark B on cpus 24-47.
This is useful for seeing the interaction of benchmarks.
I can collect perf at a system level and get IPC for the whole
system. But I'd like to see how the IPC for benchmark A changes
from when I run just benchmark A on the whole box.
Doing 'perf stat -A ...' will report the per-cpu perf event data.
So my output data dir will have something like:
archive_top
archive_top/perf data
archive_top/perlbench cpu2017 perlbench runs on half the cpus.
archive_top/omnetpp  cpu2017 omnetpp runs on the other cpuss.

I'd like to generate an xlsx for just the perlbench running on
the 1st half of the cpus and another xlsx for omne runnng on the
2nd half of the cpus. So I can get IPC for perl and the IPC for
omne.
So the way I handle this is with an option string.
file_sets{/b19a.*perlbench_r/:perf_cpus=0-23}
This matches the filenames used in the xlsx to the regex:
/b19a.*perlbench_r/ and, if I get a match then perf_cpus=0-23
is passed to perf_stat_scatter.sh
perf_stat_scatter.awk then only uses perf data from cpus 0 to 23
(which is the cpus on which perlbench is running).
This fileset stuff really should be in a json file but TBD.