Support perf_cpu_groups.txt file to select cpus
See previous commit for format of perf_cpu_groups.txt file.
Basically when perf is run you can create a perf_cpu_groups.txt
with a list of what work is running on which cpus.
This is for doing things like running cpu2017 workloads on
specific cpus.
The changes here look for perf_cpu_groups.txt in the directories
that are being used. So there has to be a perf_cpu_groups.txt file
in the used dirs. And you have to set:
perf_cpu_groups{/look_for_string/:}
option and pass it to gen_xlsx.sh.
gen_xlsx.sh passes the option to sys_2_tsv.sh.
Then sys_2_tsv.sh has to match the 'look_for_string' to a line
in the perf_cpu_groups.txt file. The look_for_string can be a
regex. The 3rd field in each line is the list of cpus for the
'look_for_string' workload. If there is a match on the string
then that list of cpus is passed to perf_stat_scatter.awk.
And then perf_stat_scatter.awk uses the list to select which
lines of perf stat data are used.