make threads/core a variable in script
Currently the threads per core is hard-coded at 2 in the script.
Make it a global variable (thr_per_core) set outside the awk
script and passed to the awk script.
There are cases where the thr_per_core needs to be 1.
In particular for data collected by process id (pid).
This commit lets me override the thr_per_core.
thr_per_core is used in the topdown equations' cycles thread_any
event. Near as I can tell, if you doing per pid data collection
then you only get the cycles for the cpu that the thread is using.
not the other cpu too.