check for the list of perf_cpu_groups cpus
Currently, if you've collected per-cpu perf data, and you haven't
gotten a match on perf_cpu_groups then the list of cpus to use
will be null. But the script will still set thr_per_core = 1.
This can lead to unexpected results, in particular if your
benchmark didn't use all the cpus.
This change doesn't set thr_per_core to 1 if the cpu_list is null.