Add push_glbl_row_arr operation and ck div by 0
Add push_glbl_row_arr operation which will pull the
glbl_row_arr[some_name] value and push it to the current equation.
This assumes that the value has been updated for the current row
already. I probably should delete the value at the top of the loop
so we can check that it has already been computed before it is
used.
Also check, when we are going to do a divide, check that the
denominator is not 0. Currently no checking is done and, if we get
a divide by 0 err, awk exits and I can't catch the error... or
at least I can't catch it in time to display things like the
equation that has the issue.