Allow -f dir_name and perf_cpu_groups
Currently the -f arg has to be a sys_*_perf_stat.txt file.
Allow arg to be a dir and then search for sys_*_perf_stat.txt file
under that dir.
Also add -c perf_cpu_groups_string to select which cpus to select
when we read the perf data file.