Add %cycles backend stalls and %cyc ret uops >= 1
The %cycles_backend_stalls is not the same as topdown_stalls but
it is the biggest component.
The %cyc_uopRet = "%cycles where 1 or more uops retired" and is a
per-thread equation. It is different than topdown %retiring.
%retiring is the % of the slots retiring uops per cycle but on a
per core level. We need something per thread if we profile
per cpu or per process (per cgroup).
So the 5cyc_uopRet eqn will tell us the number of cycles where 1
or more uops are retiring per thread.