lots of changes to create output tsv in work_dir
This script creates most of the output tsv files. Usually these
files would get created in the same dir as the source data file.
Now we want to put the output files in a work_dir.
Most of the changes are to pass work_dir to awk script and then
prefix the output tsv file with the work_dir.