exit early with error if perf_cpu_groups lkup fail
Currently if the the string you enter to select a perf_cpu_group
is not valid then the script uses the 'all' cpu group.
For example, if you are trying to select the cpus running the
omnetpp subtest and you enter 'onme' (which fails the lookup test)
then the script defaults to all cpus.
This is not what should happpen. The script should quite with an
error. Otherwise you will get cpu stats for all cpus instead of
the omnetpp cpus.
Now return an error if lookup string is not found.