do better job geting tsc frequency
Currently the script tries to get tsc freq from bogomips. This
seems to work but fails aat if HT is off. Or maybe I should just
divide bogomips by 2 always.
In any case, the right way to get it on Intel is to read the
cpu string and get the 2.2GHz string.
AMD doesn't have this string in the cpuid brand string.
In any case, I also run my 'tsc_freq.x' binary to get the tsc so
I can use that info.